% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arpabet}
\alias{arpabet}
\title{arpabet mappings}
\format{a data frame with three variables:
 \describe{
  \item{arpabet}{ARPABET symbol}
  \item{ipa}{IPA symbol}
  \item{xsampa}{X-SAMPA symbol}
}}
\usage{
arpabet
}
\description{
A dataset mapping ARPABET to IPA and X-SAMPA
}
\keyword{datasets}
