% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enGBxrpCainteoir}
\alias{enGBxrpCainteoir}
\title{Received Pronunciation Dictionary}
\format{a data frame with four variables:
 \describe{
  \item{word}{Word}
  \item{alt}{Alternate label}
  \item{pronounciation}{Pronunciation in ARPABET}
  \item{comment}{Comment}
}}
\usage{
enGBxrpCainteoir
}
\description{
A dataset of the Received Pronunciation Dictionary
}
\keyword{datasets}
