% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enScotlandCainteoir}
\alias{enScotlandCainteoir}
\title{The Scottish English Dictionary}
\format{a data frame with four variables:
 \describe{
  \item{word}{Word}
  \item{alt}{Alternate label}
  \item{pronounciation}{Pronunciation in ARPABET}
  \item{comment}{Comment}
}}
\usage{
enScotlandCainteoir
}
\description{
A dataset of the Scottish English Dictionary
}
\keyword{datasets}
