% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enUSxgaCMU}
\alias{enUSxgaCMU}
\title{CMU Prouncing Dictionary}
\format{a data frame with four variables:
 \describe{
  \item{word}{Word}
  \item{alt}{Alternate label}
  \item{pronounciation}{Pronunciation in ARPABET}
  \item{comment}{Comment}
}}
\usage{
enUSxgaCMU
}
\description{
A dataset of the CMU Pronouncing Dictionary
}
\keyword{datasets}
